// apps/server/src/infra/files-storage-client/generated/models/file-record-parent-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const FileRecordParentType = {
    USERS: 'users',
    SCHOOLS: 'schools',
    COURSES: 'courses',
    TASKS: 'tasks',
    LESSONS: 'lessons',
    SUBMISSIONS: 'submissions',
    GRADINGS: 'gradings',
    BOARDNODES: 'boardnodes',
    EXTERNALTOOLS: 'externaltools'
} as const;

export type FileRecordParentType = typeof FileRecordParentType[keyof typeof FileRecordParentType];