// apps/server/src/infra/files-storage-client/generated/models/storage-location.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const StorageLocation = {
    SCHOOL: 'school',
    INSTANCE: 'instance'
} as const;

export type StorageLocation = typeof StorageLocation[keyof typeof StorageLocation];